SqlReminderTable: Fix hash type mapping to database (issue12)
The hash is an unsigned integer in the grain, but was previously represented as
an integer in the SQL script. This caused an ArgumentException when
DbType.UInt32 was attempted to be converted to an SqlDbType as there is no
mapping from DbType.UInt32 to a corresponding SqlDbType value.

To accomodate the unsigned integer, the has in the reminder table has been
changed to a bigint, and the calling code now uses DbType.Int64 instead.