Fix for issue #492. Updated ProcessAssembly in codegenerator to verify that any Grain classes (i.e. extending from IGrain) must be either SystemTarget or must implement either Grain or Grain<>. On failure, we throw warning during code generation.
However, making this fix raises a number of warnings in the orleans
runtime codebase.
For example, IReminderTable interface is an IGrain, but it has several
implementaions such as Azure, SQL, GrainBased, out of which only
GrainBased implementation implements Grain. And hence warnings for
other classes.

One fix for this issue is to separate the IReminderTable interface into
two as follows.
interface IReminderTable {
...
}
interface IReminderTableGrain : IReminderTable, IGrain {
...
}
class AzureBasedReminderTable: IReminderTable
class GrainBasedReminderTable: IReminderTableGrain

Current codegenration does not support this and hence needs to be fixed
before this patch can be merged. (a separate PR has been submitted to enable
this feature issue # 504)