QueueCacheMissException #844 Fix
The cache cursor for the first stream is becoming inactive while there is data still being delivered on the second stream. Since data is being delivered on the second stream the cache is moving forward in time and when we try to refresh the cache cursor for the first stream, it is older than any of the data in the cache, so it throws a cache miss.

When we refresh a cursor, if it's token is no longer in the cache, we start from the oldest data in the cache.