CommandBehavior and cancellation added to relational storage
CommandBehavior support is needed to make use of the
more advanced possibilities of the underlying ADO.NET
provider, such as natively supported streaming. However,
not all ADO.NET providers support all options or there
are known "features" that are worked around as transparently
as possible.

The relational provider has not had a native
cancellation support thus far. This commit adds
it and some supporting tests. A note needs to
be made that not all ADO.NET providers support
cancellation natively and it is accounted for.

In the process some performance tweaks and
Orleans specific streaming added to classes
that do not support it natively. Also, the
return value of value selector changed to
return Task<TResult> instead of plain TResult
so that it's possible to use await when selecting
data from the database. This makes utilizing
CommandBehavior related optimizations easier.