Refactor of EventHubStreamProvider to better support plugable CacheDataAdapter.
To test the ability for users to provide their own cache data adapter to the EventHubStreamProvider, I attempted to add a test that uses a custom data adapter to produce a single stream per partition. This attempt exposed flaws in the design.

This change contains refactors to the EventHubStreamProvider to support this scenario, as well as the Stream per partition test.