ILBasedSerializer: remove direct serializers for class types since they don't handle nulls values (#2354)
Reduce the scope of an optimization used for serializing values which BinaryTokenStreamReader/Writer can read/write directly.
The issue arises when fields of those types are null, since BinaryTokenStreamReader/Writer expect that to be handled at a layer above, so NullReferenceException is thrown.
This could be handled in IL, but I don't want to make either the generator or the emitted code complex at this stage. Now only value types will be directly serialized.