Remove invalid and redundant parameter check in NewKey (#2728)
This whole check was wrong, and that it only survived because of the incorrect condition that is always false. It is totally legal for both n0 and n1 to be 0. For example, the very next overload NewKey(long longKey, does

var n1 = unchecked((ulong)longKey);
return NewKey(0, n1, category, typeData, keyExt);
where longKey can be zero.