Add support for minimal disambiguation of service types over DependencyInjection.Abstractions (#2843)
Since we're limited to features provided by Microsoft.Extensions.DependencyInjection.Abstractions, neither we nor our users that don't use more advanced DI, have a means of acquiring a service from a service provider by name or key.  This PR introduces a minimal implementation for this which we'll use in several efforts in parallel.