Orleans Facet System initial infrastructure (#3279)
Infrastructure for "Orleans Facet System #3223", and a test feature "ExampleStorage" to vet the intended pattern.

**Infrastructure**
The facet system infrastructure consists of a utility class (ConstructorArgumentFactory) which constructs facet arguments for grains, and an extensible pattern for mapping attributes to cacheable feature factories (IAttributeToFactoryMapper). This is syntactic sugar to avoid the direct use of a factory object to get to the underlying feature.

**ExampleStorage feature**
The test feature ExampleStorage explored one of our more commonly expected feature patterns.  It is a generic service, selectable by name, which requires grain specific configuration.  In this sample implementation we break the feature in to three areas.

- Abstractions - We expose the feature interface, expected configuration, factories, and attribution in the abstractions layer.  Grains should only need to access these abstractions.
- Infrastructure - The feature infrastructure deals with the feature factories and how the attribute is wired to the factories.  In the case of the ExampleStorage, we do this by calling the feature factory which leverages the existing names services capability to select a factory by name, and build the instance with the factory.
- Implementations - Each implementation of the ExampleStorage is independently registered as a named service, making it available to the ExampleStorage infrastructure.

**Developer Audiences**
The separation of the test ExampleStorage into abstractions, infrastructure, and implementations, illustrates the following prioritization:

1) Grain developer experience- Via the clean abstraction layer, grain developers need not concern themselves with implementation details.
2) Feature extensions - Extending the feature (adding support for new types of storage) is supported by the named service registration capability.  Developers adding new implementations of a feature need only implement the feature and it's factories, and register the factory as a names service.
3) New feature developers - Most of the work is placed on the feature developer to set all of this up.  While adding new features takes the most work, having this pattern still makes it easier and more straight forward to do, and exposes the features in a way that is very grain developer friendly.

This prioritization is intentional as it is expected that most developers will use features, many may extend features, and far fewer will add new features.