Add deactivation injection transaction tests (#4845)
- implemented DeactivationTransactionalState facet which can inject deactivation command into different phase of a transaction
- implemented DeactivationTransactionTestRunner and AzureDeactivationTransactionTests which uses the runner. This runner contains a minimum scenario which can happen during our recovery tests

The goal of this test runner is to break down transaction recovery tests into more granular test cases. Our current recovery test is a end to end test, we shutdown a silo (non)gracefully during the test, and see if transaction recovered. It is failing now, and it is very hard to debug since test may fail due to a different reason each time, and there's a broad range of possibilities to explore. This test runner introduced a list of possible scenarios which can happen in current recovery tests. We can add more possible scenarios later after this pr. 

Besides the value of fixing recovery tests, this test suit also should exist to ensure certain scenario works.