Pass PlacementTarget into Filter method (#9482)
pass PlacementTarget in to Filter method

This allows for Request Context to be used for placement decisions.

In the future we may want to allow for caching of filtering decisions (between membership updates and potentially some other criteria that we get from the implementation of a filter), but that would need to be opt-in from the filter implementation and the filter would need to accept that it's results may be cached. This would still be possible after this change as it would rely on the filter implementation's intent to be cached and not strictly enforced (this is a similar pattern to `[Immutable]`)