[Azure] Package version bumps (#9453)
* [Azure] Package version bumps

The focus of these changes is to bump Azure SDK
package versions.  Of particular note is Event Hubs, which
absorbed a service contract change that requires
offsets to be treated as strings, as the format is
no longer guaranteed to be numeric.

In response to this, uses of offset-related members
were shifted to the "OffsetString" version.  In
addition, I removed some use of deprecated protected
constructors in favor of the model factory for
creating instances with overrides for broker-owned
fields.

* Marking obsolete exception serialization constructor as obsolete and non-browsable.

* Update Directory.Packages.props

* Bumping logging abstractions to avoid downgrade